pyproject.toml
[tool.poetry]
name = "iris"
version = "1.0.0"
description = "Universal Visual Target Detection System"
authors = ["Development Team <dev@company.com>"]
readme = "README.md"
packages = [{include = "modules"}]

[tool.poetry.dependencies]
python = "^3.9"
mindar = "0.1.2"
opencv-python = "^4.5.0"
paho-mqtt = "^1.6.0"
PyYAML = "^6.0"
numpy = "^1.26.0"
msgpack = "^1.0.0"
numba = "^0.59.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
ar-detector = "ar_target_detector:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
